+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Please mark all changes in change log and use the issue from GitHub
 -   \#2649 Search parameter of annoy has conflict with document
 -   \#2690 Remove body parser in show-partitions endpoints
 -   \#2692 Milvus hangs during multi-thread concurrent search
+-   \#2739 Fix mishards start failed
+-   \#2776 Fix too many data copies during creating IVF index
 
 ## Feature
 -   \#2319 Redo metadata to support MVCC